Add note regarding `build.rs` to MSRV in `Cargo.toml` (#17)
## Description

This adjusts the comments regarding the Rust version used when invoking
`bindgen` in `build.rs`. The version there should be more or less kept
in sync with the MSRV in `Cargo.toml`.

We say "more or less" because bindgen does only track Rust versions that
affect the ABI. When we have an MSRV that had no ABI changes, we leave
the version in `build.rs` unchanged at the previous version (we must not
use any newer version).