test: Add full client/server test (#72)
This is a side effect of upgrading to open62541
[master](https://github.com/open62541/open62541/tree/master) (to be 1.5
eventually): we now run an integration test that spawns up a server,
then uses the client to connect to it and read a value.

This requires an upgrade to MSRV 1.82+. We use 1.83 to stay in line with
[open62541](https://github.com/HMIProject/open62541). This also upgrades
all CI workflows as a follow-up to #70.